feat: 인가 추가, 코스 도메인에 좋아요 도메인 합침
